+++ pynaja/__init__.py
@@ -1,3 +1,3 @@
 # -*- coding: utf-8 -*-
 
-__version__ = r'1.0.1'
+__version__ = r'1.0.2'
